feat: AbstractHttpRequestHandler를 통한 handleHttpRequest 중복 제거
